docs: add tutorial on data and model drift (#4652)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Add a tutorial on which we simulate a real scenario where we have
fine-tuned a model on historical data and observe a degradation in the
accuracy using this model. Then we analyze data to understand why data
drift happened.

TO DO:
- [x] Add more info on some parts
- [x] Rerun some outputs

NOTES:
-I used Ollama as not all people usually have access to an openai
api.key, and we're using it in most of the tutorials.
-To analyze data drift I provide two options so that they don't have to
install everything. I also included text descriptions, as in many
articles, were mentioned.

Closes #4448 

**Type of change**

(Remember to title the PR according to the type of change)

- [x] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes.)

- [x] `sphinx-autobuild` (read [Developer
Documentation](https://docs.argilla.io/en/latest/community/developer_docs.html#building-the-documentation)
for more details)

**Checklist**

- [ ] I added relevant documentation
- [x] I followed the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)