refactor: improve NotFoundError exceptions raised on API v1 and how records are retrieved from database (#4876)
# Description

This is the first of some PRs refactoring and improving the error
handling for API v1.

In this specific PR I'm improving how `NotFoundError` is managed in the
API v1 with the following changes:
* Now `NotFoundError` is added to the list of the captures exceptions
for API v1.
* The response for `NotFoundError` is not changed using the same
structure: `{"detail": "message"}`. Once we are working on v2.0 we can
change it to use the new one.
* I have added two new methods to our database models so the retrieval
of instances is simplified:
  * `get`: useful to get an instance of a database model using an id.
* `get_or_raise`: useful to get an instance of a database model using an
id and raising a `NotFoundError` exception if the instance is not found.
* The two methods include a `options` parameter that can be used to
pre-load associations.

Refs #4871 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Changing tests and adding new ones when necessary.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)