refactor: add exception handler for ValueError exceptions (#4913)
# Description

This PR add an exception handler for `ValueError` exceptions responding
with a 422 and a JSON structure like `{"detail": "..."}`.

We added this exception handler to avoid problems raising `ValueError`
exceptions and returning a 500 error instead of a 422.

Once that we stop using `ValueError` and use `UnprocessableEntityError`
exception instead we can remove this exception handler.

Refs #4871 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Modifying some test to raise a `ValueError` and checking that the
exception is correctly captured.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)