refactor: add `get_by` and `get_by_or_raise` class methods to database models (#4919)
# Description

This PR adds the following changes:
* Add a new `get_by` class method which will return an instance of a
model using a set of filter/conditions. If the instance is not found it
will return `None`.
* Add a new `get_by_or_raise` class method which will return an instance
of a model using a set of filter/conditions. If the instance is not
found it will raise a `NotFoundError`.
* Remove `read` and `read_by` unused class methods.

Apart from these changes this PR doesn't change the behavior of the API
except some error messages when an instance is not found using
`get_by_or_raise` where the string message of the error is improved to
show the parameters.

Closes #4916, #4918 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Modifying some tests and checking that everything is working as
expected.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)