refactor: move and improve API policies structure (#4929)
# Description

This is the first PR of a series of PRs improving the folder structure
for our API.

For this one I'm moving and making some improvements to the API policies
without changing the behavior.

The new folder structure for the policies in this PR is the following:
```
/api
  /policies
    /v1
       model_a_policy.py
       model_b_policy.py
       ...
```

Notice how the policies are now moved to individual files instead of one
unique big file.

I have also made some improvements to the logic checking the permissions
inside the policy and created a `User.is_member` function. This function
is the first one doing database queries inside a database model but I
think it makes sense to have these logic here. For more information
about this take a look to this commit:
https://github.com/argilla-io/argilla/commit/b4111bdcc08762492bee9dd1f58cb61e82d8a1be

Ref #4868

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] We are not modifying behavior so tests should be passing after the
changes.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)