bug: check search engine connection on application startup (#4970)
# Description

The connection testing on application startup for search engine was
removed on this
[change](https://github.com/argilla-io/argilla/pull/4853/files#diff-a6a4ff8899f30f559bc6361447bfeb449545bff039586f4fed91241556755cc6)
because it was using code that was removed for v2.0.

In this PR I'm adding a ping request to the search engine using the
`SearchEngine` class. The search engine client already has a backoff
mechanism retrying for 6 times incrementing the waiting time. If the
instance is not responding after all the retries the ping will return
`false`, raising and exception and finally closing the app.

Btw I'm using branch about improving structure as base because I was
suffering a import error that it's solved in these PRs.

Closes #4946 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Manually checking that the behavior is the expected one with
ElasticSearch up and down.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)