docs: add changelog (#4983)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This is the preview. The changelog generation doesn't affect the
building time. Maybe it would be better to remove the right toc.

![Captura de pantalla 2024-06-09
125213](https://github.com/argilla-io/argilla/assets/127759186/80158924-6d47-4d3c-a4c2-17b99c357db6)

Closes #4972 

**Type of change**

(Remember to title the PR according to the type of change)

- [ ] Documentation update

**How Has This Been Tested**

```
mkdocs serve
```

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>