docs: popular issues file generator (#4971)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Script to automatically retrieve the github issues and generate a doc
with the most popular.

NOTE: By the moment, I left it commented in mkdocs.yml as I don't know
if this will be added for 2.0. So, when merged, it's saved but not run.
Using it, the docs building takes around 20 seconds more.

TO BE DISCUSSED (below a screenshot):

1. Still not sure about the given names to each table, probably it's
better to shorten them.
2. I added it to Community, but if you think this is more suitable in
other section, we can discuss it.
3. I ran it locally, but I don't know if any external workflow must be
also updated.
4. Any other feedback is welcome!


![screencapture-127-0-0-1-8000-argilla-python-community-popular-issues-2024-06-06-14_32_12](https://github.com/argilla-io/argilla/assets/127759186/660656f9-47b9-4703-8411-ef4f2389d068)


Closes #4967 

**Type of change**

(Remember to title the PR according to the type of change)

- [x] Documentation update

**How Has This Been Tested**

- [x] Run with `mkdocs serve`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: David Berenstein <david.m.berenstein@gmail.com>