chore: add new argilla-server folder structure to README.md (#4976)
# Description

In this PR I'm adding a small section to our `README.md` file with the
new folder structure implemented.

@frascuchon I have checked too if it was possible to move `/security`
folder to `/api` new folder structure but I think it should not be part
of `/api`. In any case it should be moved to `/context` not sure if it's
ok to move it right now as it is. (I have removed one file inside
`/security` that was not used at all)

Refs #4868

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [x] Documentation update

**How Has This Been Tested**

- [x] Running tests.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)